--- data/park.txt
@@ -1,1 +1,0 @@
-T/[ ]/amogus
