+++ src/main/java/park/Park.java
@@ -1,14 +1,11 @@
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.time.format.DateTimeParseException;
-import java.util.Date;
-import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.List;
-import java.io.File;
-import java.io.FileWriter;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+package main.java.park;
+
+import main.java.park.commands.Command;
+import main.java.park.exceptions.ParkException;
+import main.java.park.parser.Parser;
+import main.java.park.storage.Storage;
+import main.java.park.storage.TaskList;
+import main.java.park.ui.Ui;
 
 public class Park {
