+++ src/main/java/park/commands/ExitCommand.java
@@ -1,3 +1,9 @@
+package main.java.park.commands;
+
+import main.java.park.storage.Storage;
+import main.java.park.storage.TaskList;
+import main.java.park.ui.Ui;
+
 public class ExitCommand extends Command{
 
     @Override
