+++ src/main/java/park/exceptions/ParkException.java
@@ -1,3 +1,5 @@
+package main.java.park.exceptions;
+
 public class ParkException extends Exception {
 
     public ParkException() {}
