+++ src/main/java/park/task/ToDo.java
@@ -1,4 +1,6 @@
-public class ToDo extends Task{
+package main.java.park.task;
+
+public class ToDo extends Task {
 
     public ToDo(String description) {
         super(description);
